Migrate to Typescript
