+++ .gitignore
@@ -1,2 +1,4 @@
+/lib
 /node_modules
+yarn-error.log
 package-lock.json
