+++ .npmignore
@@ -1,0 +1,8 @@
+.circleci
+/test
+/src
+tsconfig.json
+.editorconfig
+Dockerfile
+now.json
+LICENSE
