+++ package.json
@@ -19,20 +19,29 @@
     "url": "git+https://github.com/zeit/dns-cached-resolve.git"
   },
   "scripts": {
-    "fmt": "prettier --single-quote --print-width=80 --write index.js",
+    "prepublishOnly": "yarn build",
+    "fmt": "prettier --single-quote --print-width=80 --write './{src, test}/**/*.ts'",
+    "build": "tsc",
     "test": "jest"
   },
-  "main": "index.js",
+  "main": "lib/index.js",
   "dependencies": {
+    "@types/async-retry": "^1.2.1",
+    "@types/lru-cache": "^4.1.1",
+    "@types/node": "^10.12.18",
     "async-retry": "1.1.3",
     "lru-cache": "4.1.1"
   },
   "devDependencies": {
+    "@types/jest": "^23.3.10",
     "jest": "23.4.2",
-    "prettier": "1.14.2"
+    "prettier": "1.14.2",
+    "ts-jest": "^23.10.5",
+    "typescript": "^3.2.2"
   },
   "jest": {
-    "verbose": true,
-    "testURL": "http://localhost/"
+    "preset": "ts-jest",
+    "testURL": "http://localhost/",
+    "verbose": false
   }
 }
